Pull more nav stuff into side nav
